+++ src/test/java/org/folio/search/service/reindex/jdbc/CallNumberRepositoryIT.java
@@ -109,10 +109,10 @@ class CallNumberRepositoryIT {
       .contains(
         tuple("number1",
           List.of(mapOf("instanceId", List.of("9f8febd1-e96c-46c4-a5f4-84a45cc499a2"),
-            "locationId", null, "shared", false, "tenantId", TENANT_ID))),
+            "count", 1, "locationId", null, "shared", false, "tenantId", TENANT_ID))),
         tuple("number2",
           List.of(mapOf("instanceId", List.of("9f8febd1-e96c-46c4-a5f4-84a45cc499a2"),
-            "locationId", null, "shared", false, "tenantId", TENANT_ID))));
+            "count", 1, "locationId", null, "shared", false, "tenantId", TENANT_ID))));
   }
 
   @Test